update SEConsent & specs
